Updates sync action to use a stable branch
The current sync action implementation opens a new PR every time there
is a new commit on master. This implementation will use a stable branch
when it opens PRs so that the branch can be updated on subsequent
changes and the PR will get updated if it is still open.