add action that adds an issue to project and set fields
This action adds the given issue to a github project board and
(optionally) sets a custom field of type Single-Select to the given
value.
(e.g. in the Paketo Workstreams board, adding an issue and setting its
"Workstream" field) This action is designed to be chained with
`actions/issue/file`.

Here's a typical example of usage:
```yaml
    - name: File Issue
      id: file-issue
      uses: paketo-buildpacks/github-config/actions/issue/file@main
      with:
        token: <token>
        repo: ${{ github.repository }}
        issue_title: "Some title"
        issue_body: |
          Some issue body

    - name: Add issue to project
      id: issue-to-proj
      uses: paketo-buildpacks/github-config/actions/issue/add-to-project@action-issue-add-to-proj
      with:
        token: <token>
        issue-node-id: ${{ steps.file-issue.outputs.node-id }}
        project-org: my-org
        project-num: 24
        field-name: Workstream
        option-name: Third Workstream
```

Github CLI does not currently allow interacting with the API of new-style projects.
GraphQL api doc:
https://docs.github.com/en/issues/planning-and-tracking-with-projects/automating-your-project/using-the-api-to-manage-projects

Related: https://github.com/paketo-buildpacks/github-config/pull/665