Use go version from go.mod file (#1146)
* Use go version from go.mod file

Use the go version from the project's go.mod file, not a hard coded version. This should enable the project to set the Go version it requires and avoid situations where there are CI failures because the CI go version was updated before the project was ready.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>

* Too aggressive, this repo doesn't have a go.mod file

Signed-off-by: Daniel Mikusa <dan@mikusa.com>

* Ensure checkout happens before setup-go so go.mod file is present

Increase space between comment to fix yaml lint errors

---------

Signed-off-by: Daniel Mikusa <dan@mikusa.com>
Co-authored-by: Jerico Pena <jericop@gmail.com>
Co-authored-by: Jerico Pena <jericop@users.noreply.github.com>