feat(integration): add action to free space on runner (#1237)
* feat(integration): add action to free space on runner

Some integration tests, more specifically those for
python, often fail with an out of space error.

This action frees some space on the runner which allows
these tests to pass successfully.

* refactor: integrate implementation from jlumbroso/free-disk-space PR 48

The original action has not seen much movement lately
and the PR adds good improvements.

Keeping the code in this repository removes the dependency.

* fix: correct linter detected issues

* fix: it's vars and not env that must be used

The variables sets through GitHub repository settings
are stored in vars (which can be used to fill env).

Details:
https://docs.github.com/en/actions/reference/workflows-and-actions/contexts#vars-context